+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Report `LanguageServerInstallationStatus::Failed` to Zed in case of errors in the install process.
 - Add some supported languages to the extension manifest.
 - Add a GitHub Actions workflow to publish LSP server binaries to GitHub Releases on new version tags.